% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHclean_script.R
\name{HHclean}
\alias{HHclean}
\title{Transforming rRNA/ITS sequencing data to a similar, standard format}
\usage{
HHclean(df, year)
}
\arguments{
\item{df}{The name of the specified data set}

\item{year}{The year the that the data set was collected}
}
\description{
This function takes the an input of sequenced rRNA/ITS data and changes its structure from wide, to long. It also categorizes the information by the phylum taxonomy, and adds a column to indicate the year that the data samples were taken.
}
