% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHplot_script.R
\name{HHplot}
\alias{HHplot}
\title{Plotting stacked bar charts of abundance data}
\usage{
HHplot(df1, df2 = NULL)
}
\arguments{
\item{df1}{The name of one specified data set}

\item{df2}{The name of the second specified data set. If NULL, the default, the function will just use the first specified dataset to plot a stacked bar chart.}
}
\description{
This function takes cleaned datasets as an input. If there is only one clean dataset input, it will plot a stacked bar chart. If there are two clean dataset inputs, it will plot a clustered stacked bar chart. It allows for comparison through different years that data is collected.
}
