% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HHsplit_script.R
\name{HHsplit}
\alias{HHsplit}
\title{Split and specify a taxonomic family}
\usage{
HHsplit(df)
}
\arguments{
\item{df}{The name of the specified data set}
}
\description{
This function takes raw rRNA/ITS sequencing data frame as the input. The function takes the entire taxonomic name of a bacterial unit and splits it by the semicolons. The semicolons divide each of the different taxonomies. After it splits the complete taxonomic name, the function specifies and pulls out only the species' phylum level.
}
